Relative number, solarized, always show tab line
Using a simple plugin that will remove relative number when in insert
mode.

Updated theme to solarized light so it's easier to display when giving a
talk or a workshop.

Always show status line even if only one thing is open. This is useful
since the tab is the easiest way to see if you saved the file at a
glance or not. But if you don't open a new tab you wont see the tab
line. Now that is fixed.

Updated a few comments.