Add Tmuxinator config to dotfiles
The beauty of tmux is that you can manage multiple projects and switch
between them quickly. Tmuxinator is a tool I use to quickly setup a
project so I can restart a project right away, the way I want it.

I want Tmuxinator to put the project configs inside dotfiles local.

I don't want project configs that are specific to my local environment
added to version control. But dotfiles-local is always the same so I'll
add that.