Tmux Quality of life
I added a function and alias called "tma" which will do a few things.
When provided with a session name ("tma my-project") it will first try
to attach to an existing tmux session with that name. If it doesn't
exist, it will ask Tmuxinator to bootstrap a session with that name. If
there isn't an existing config for that session name, then create a new
regular tmux session with that name. Easy peasy. Saves a lot of typing.